[Refactor] Move app start code to class
